Updated Search, Visuals, delete function
Added product_Type to search so if user were to type "ram" it will show the ram types and if ram showed up anywhere else. Updated parts visual by changing color of price, and adding an if statement to print if a product is in stock (green) or not (red). Fixed bug of not showing image of products in cart. Updated account page to confirm that the username = "admin" cannot delete their own account. Another type "root_admin" however can delete their account.